refactor(libtau): replace Disk backend with Sstable, remove disk_decode fuzzer
Sstable is now the sole on-disk backend everywhere (server config,
executor,
DST matrix, fuzzing) rather than living alongside the old
whole-file-rewrite
Disk format; manifest version resets to v1 since nothing shipped yet.
Docs
across crates/ and docs/ updated to match
