+++ .gitignore
@@ -12,4 +12,3 @@ crates/fuzztau/corpus
 /crates/fuzztau/target
 /demo
 /sample_data
-PLAN.md
