+++ crates/README.md
@@ -7,7 +7,7 @@ Tau is a Cargo workspace. One engine library, two binaries, and one fuzz harness
 | [`libtau`](libtau/src/README.md) | library | The engine: data model, query language, storage backends, executor. Everything else depends on it. |
 | [`tau`](tau/README.md) | binary | TCP server: exposes a `libtau` executor over a line-oriented protocol with optional TLS, auth, and WAL. |
 | [`tauctl`](tauctl/README.md) | binary (`tauctl`) | Interactive client: named connection pool, lazygit-style pane navigation, clipboard copy/paste, client-side CSV load. |
-| [`fuzztau`](fuzztau/README.md) | fuzz harness | LibFuzzer targets for the TauQL/wire text parsers and the binary `.dat` decoder (`disk_decode`); a TauQL dictionary and committed seeds; requires a nightly toolchain. |
+| [`fuzztau`](fuzztau/README.md) | fuzz harness | LibFuzzer targets for the TauQL/wire text parsers; a TauQL dictionary and committed seeds; requires a nightly toolchain. |
 
 Binary names are unique across the workspace, so `cargo run --bin tau|tauctl` works from the repo root.
