+++ crates/dst/src/profile.rs
@@ -75,10 +75,6 @@ impl ProfileSpec {
                 let d = paths.disk_dir.as_ref().expect("disk dir");
                 bootstrap_disk(d, self.compact_threshold(), self.enc_key())
             }
-            Storage::Sstable => {
-                let d = paths.disk_dir.as_ref().expect("disk dir");
-                bootstrap_sstable(d, self.compact_threshold(), self.enc_key())
-            }
         }
     }
 
@@ -87,8 +83,9 @@ impl ProfileSpec {
         bootstrap_oracle_model(self, paths)
     }
 
-    /// Reopen the disk-backed executor from its existing `.dat` files without
-    /// re-issuing DDL — models a process restart for the persistence checks.
+    /// Reopen the disk-backed executor from its existing manifest/run files
+    /// without re-issuing DDL — models a process restart for the persistence
+    /// checks.
     #[cfg(test)]
     pub fn reopen_disk_executor(&self, paths: &ProfilePaths) -> Executor {
         let d = paths.disk_dir.as_ref().expect("disk dir");
@@ -185,34 +182,10 @@ fn bootstrap_disk(dir: &Path, threshold: usize, enc_key: Option<[u8; 32]>) -> Ex
     ex
 }
 
-fn bootstrap_sstable(dir: &Path, threshold: usize, enc_key: Option<[u8; 32]>) -> Executor {
-    let mut ex = Executor::with_sstable_backend(
-        dir,
-        threshold,
-        libtau::storage::DEFAULT_ZSTD_LEVEL,
-        enc_key,
-        true,
-        None,
-    )
-    .expect("sstable backend open");
-    exec(&mut ex, "CREATE DATABASE default");
-    for lens in INT {
-        exec(&mut ex, &format!("CREATE LENS {lens} int"));
-    }
-    exec(&mut ex, &format!("CREATE LENS {FL} float"));
-    exec(&mut ex, &format!("CREATE LENS {BOOL} bool"));
-    exec(&mut ex, &format!("CREATE LENS {SV} str"));
-    exec(&mut ex, "CREATE DATABASE aux");
-    exec(&mut ex, "USE DATABASE aux");
-    exec(&mut ex, &format!("CREATE LENS {} int", Lens::Aux.as_str()));
-    exec(&mut ex, "USE DATABASE default");
-    ex
-}
-
 /// Reopen an existing disk-backed executor **without** re-issuing schema DDL —
-/// `CREATE DATABASE` replays each `.dat` file's persisted schema, so lenses and
-/// policies come back automatically. Models a real process restart (used by the
-/// disk-persistence DST coverage).
+/// `CREATE DATABASE` replays each database's persisted manifest schema, so
+/// lenses and policies come back automatically. Models a real process
+/// restart (used by the disk-persistence DST coverage).
 #[cfg(test)]
 pub fn reopen_disk(dir: &Path, threshold: usize, enc_key: Option<[u8; 32]>) -> Executor {
     let mut ex = Executor::with_disk_backend(
