+++ crates/dst/src/profile/spec.rs
@@ -13,10 +13,9 @@ pub enum SuiteTier {
 pub enum Storage {
     Memory,
     Wal,
-    Disk,
-    /// On-disk SSTable + read-time MVCC backend (immutable runs + manifest).
+    /// On-disk `Sstable` backend (immutable runs + manifest, read-time MVCC).
     /// See `libtau::storage::Sstable`.
-    Sstable,
+    Disk,
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -79,8 +78,7 @@ impl ProfileSpec {
     }
 
     pub const fn uses_disk_dir(self) -> bool {
-        matches!(self.storage, Storage::Disk | Storage::Sstable)
-            && matches!(self.transport, Transport::Direct)
+        matches!(self.storage, Storage::Disk) && matches!(self.transport, Transport::Direct)
     }
 
     pub const fn is_wire(self) -> bool {
@@ -96,7 +94,6 @@ impl ProfileSpec {
             Storage::Memory => "memory",
             Storage::Wal => "wal",
             Storage::Disk => "disk",
-            Storage::Sstable => "sstable",
         };
         let compact = match self.compaction {
             Compaction::Default => "default",
@@ -125,19 +122,14 @@ impl ProfileSpec {
         } else {
             match self.storage {
                 Storage::Memory => 2000,
-                Storage::Wal | Storage::Disk | Storage::Sstable => 500,
+                Storage::Wal | Storage::Disk => 500,
             }
         }
     }
 
     pub fn engine_matrix(tier: SuiteTier) -> Vec<ProfileSpec> {
         let mut out = Vec::new();
-        for &storage in &[
-            Storage::Memory,
-            Storage::Wal,
-            Storage::Disk,
-            Storage::Sstable,
-        ] {
+        for &storage in &[Storage::Memory, Storage::Wal, Storage::Disk] {
             for &compaction in &[Compaction::Default, Compaction::Stress] {
                 for &encryption in &[Encryption::Plain, Encryption::Aes256] {
                     if matches!(storage, Storage::Memory)
@@ -245,12 +237,6 @@ impl ProfileSpec {
                 Encryption::Plain,
                 Transport::Direct,
                 Auth::Off,
-            ) | (
-                Storage::Sstable,
-                Compaction::Default,
-                Encryption::Plain,
-                Transport::Direct,
-                Auth::Off,
             )
         )
     }
@@ -261,8 +247,8 @@ mod tests {
     use super::*;
 
     #[test]
-    fn smoke_has_six_cells() {
-        assert_eq!(ProfileSpec::engine_matrix(SuiteTier::Smoke).len(), 6);
+    fn smoke_has_five_cells() {
+        assert_eq!(ProfileSpec::engine_matrix(SuiteTier::Smoke).len(), 5);
     }
 
     #[test]
@@ -273,7 +259,7 @@ mod tests {
                 .iter()
                 .any(|p| { p.storage == Storage::Wal && p.encryption == Encryption::Aes256 })
         );
-        assert_eq!(specs.len(), 14);
+        assert_eq!(specs.len(), 10);
     }
 
     #[test]
