+++ crates/libtau/src/storage/README.md
@@ -31,11 +31,7 @@ so a backend only needs to override them for a real performance win (`Sstable` d
 
 **`InMemory<V>`** — `FxHashMap<String, Vec<Layer<V>>>`. Zero I/O. Used by all unit tests, and the in-memory server mode. Compaction fires automatically when a lens exceeds the threshold.
 
-**`Disk<V>`** — binary flat file (format version 2) with zstd block compression. Header is the 4-byte magic `TAUZ` + VERSION + FLAGS byte + CRC32 (covers magic/version/flags). FLAGS bit 0 signals AES-256-GCM encryption; the compressed body is encrypted when set (compress-then-encrypt order). The body is a schema section (length-prefixed DDL strings) followed by the layer entries. All writes go through `flush()`, which serialises schema + entries, compresses at the configured level (default 3), optionally encrypts, and atomically replaces the file via a `.tmp` rename. A flush fires on **every append** and on every `append_schema`, so both data and DDL (`CREATE LENS` / `DERIVE LENS` / `XDERIVE LENS` / `SET TTL` / `DROP LENS`) are durable before the call returns and survive a restart — the executor replays the schema when `CREATE DATABASE` re-opens the file. Adjust the compression trade-off with `set_compression_level(level)` (1 = fastest, 22 = best ratio); exposed via `[disk] compression_level` in `config.toml`. Version 1 files (no schema section) still open.
-
-> **Note:** the server backend is configurable via `[disk] backend = "disk"` in `config.toml` (default: `"memory"`).
-
-**`Sstable<V>`** — on-disk SSTable + read-time MVCC backend: a memtable (same `Arc<[Layer]>` RCU shape as `InMemory`) flushes to immutable, zstd-compressed run files on checkpoint, listed by a small atomically-rewritten manifest. Reads merge the memtable with runs and resolve newest-wins/`AS OF` at query time; a run is skipped without decoding when its footer (per-lens min/max + a range-bucketed bloom filter) proves it can't cover the query, and a decoded run body is cached (immutable once written) so repeat queries don't re-decompress it. Compaction has exactly one trigger — see `sstable.rs`'s module doc for why that matters and what replaced the earlier three-trigger design. `DROP LENS` bumps a per-lens epoch (persisted in the manifest) so pre-drop run data is shadowed without rewriting old files.
+**`Sstable<V>`** — the on-disk backend, configurable via `[disk] backend = "disk"` in `config.toml` (default: `"memory"`): a memtable (same `Arc<[Layer]>` RCU shape as `InMemory`) flushes to immutable, zstd-compressed run files (`TAUR` magic) on checkpoint instead of rewriting anything already on disk, listed by a small atomically-rewritten manifest (`TAUM` magic). Reads merge the memtable with runs and resolve newest-wins/`AS OF` at query time; a run is skipped without decoding when its footer (per-lens min/max + a range-bucketed bloom filter) proves it can't cover the query, and a decoded run body is cached (immutable once written) so repeat queries don't re-decompress it. Compaction has exactly one trigger — see `sstable.rs`'s module doc for why that matters and what replaced the earlier three-trigger design. `DROP LENS` bumps a per-lens epoch (persisted in the manifest) so pre-drop run data is shadowed without rewriting old files. Adjust the compression trade-off with `set_compression_level(level)` (1 = fastest, 22 = best ratio); exposed via `[disk] compression_level` in `config.toml`. Encryption is AES-256-GCM (compress-then-encrypt), keyed by `TAU_ENCRYPTION_KEY`, applied separately to each run's body and footer.
 
 **`Wal`** — write-ahead log. Two entry kinds: data entries (CRC32-prefixed binary) and schema entries (`S:` / `SE:` prefix carrying raw lens DDL — `CREATE LENS`, `DERIVE LENS`, `XDERIVE LENS`, `SET TTL`, `UNSET TTL`, `DROP LENS`). Schema entries are replayed separately from data on startup. Set `set_fsync_each(false)` and call `sync()` periodically for group-commit mode.
 
@@ -55,4 +51,4 @@ both backends; revisit only if a real workload shows layers with thousands of ge
 
 `Database<V>` owns a `Box<dyn Store<V>>` behind `Arc<RwLock<>>` and an optional `Wal` behind `Arc<Mutex<>>`. Append order is WAL-first then store; a WAL fsync failure leaves the in-memory state unchanged.
 
-`Database::layers()` returns `Option<Arc<Vec<Layer<V>>>>` — an Arc-wrapped snapshot so range query phases share one allocation without re-locking.
+`Database::layers()` returns `Option<Arc<[Layer<V>]>>` — an Arc-wrapped snapshot so range query phases share one allocation without re-locking.
