+++ docs/content/docs/testing.md
@@ -100,15 +100,17 @@ Fuzz targets exercise the untrusted-input surfaces:
 | `value_decode` | `libtau::Value::decode` | UTF-8 | Value codec used inside `VAL` / `RANGE` wire segments (escapes, tags) |
 | `perm_parse` | `libtau::Perm::parse` | UTF-8 | Permission bitmap parser (`CRUDA`, `*`, `-`) used by wire `GRANTS` and users file |
 | `parse_literal` | `libtau::parse_literal` | UTF-8 | Single-literal parser used by bulk/COPY paths |
-| `disk_decode` | `Disk::<Value>::decode_payload_bytes` / `decode_image_bytes` | **raw bytes** | **Binary** on-disk `.dat` deserialization — the corrupted-file path |
 
-The text targets gate on `&str`; `disk_decode` is the only one taking **raw bytes**, exercising the binary deserialization layer where corruption desyncs length prefixes and interval bounds. It found two real bugs on its first run — an OOM from an unbounded `Value::Str` length and a panic on overlapping decoded taus — both fixed and pinned by `seeds/disk_decode/regression-*`. (DST's disk/WAL corruption faults earlier caught the inverted-interval panic in the same layer; fuzzing and DST are complementary here.)
+All current targets gate on `&str`. There is no binary-format target today — the earlier
+`disk_decode` target, which fuzzed the now-removed `Disk` backend's `.dat` decoder, was deleted
+when `Disk` was replaced by the `Sstable` backend; an equivalent target against `Sstable`'s
+run/manifest decode path would be worthwhile follow-up work.
 
 **Dictionary:** `crates/fuzztau/tauql.dict` lists TauQL/wire tokens; pass it to the text targets with `-dict=crates/fuzztau/tauql.dict` for faster structural coverage.
 
-**Seed corpus:** `crates/fuzztau/seeds/{parse,wire,value_decode,perm_parse,parse_literal,disk_decode}/` — small committed sets of valid + boundary cases (plus regression reproducers). The working corpus grows in the gitignored `crates/fuzztau/corpus/` and can be minimised with `cargo fuzz cmin`.
+**Seed corpus:** `crates/fuzztau/seeds/{parse,wire,value_decode,perm_parse,parse_literal}/` — small committed sets of valid + boundary cases. The working corpus grows in the gitignored `crates/fuzztau/corpus/` and can be minimised with `cargo fuzz cmin`.
 
-**CI:** the `fuzz` job builds every target then runs `parse`, `wire`, and `disk_decode` time-boxed (45 s each) with the dictionary and seeds; a crash/panic/OOM fails the job and uploads the reproducer.
+**CI:** the `fuzz` job builds every target then runs `parse` and `wire` time-boxed (45 s each) with the dictionary and seeds; a crash/panic/OOM fails the job and uploads the reproducer.
 
 **Prerequisites:** a nightly Rust toolchain (`rustup toolchain install nightly`).
 
@@ -126,7 +128,6 @@ cargo +nightly fuzz run --fuzz-dir crates/fuzztau wire crates/fuzztau/corpus/wir
 cargo +nightly fuzz run --fuzz-dir crates/fuzztau value_decode crates/fuzztau/corpus/value_decode
 cargo +nightly fuzz run --fuzz-dir crates/fuzztau perm_parse crates/fuzztau/corpus/perm_parse
 cargo +nightly fuzz run --fuzz-dir crates/fuzztau parse_literal crates/fuzztau/corpus/parse_literal
-cargo +nightly fuzz run --fuzz-dir crates/fuzztau disk_decode crates/fuzztau/seeds/disk_decode -- -rss_limit_mb=2048
 
 # Minimise a corpus after a long run (keeps high coverage, drops redundant inputs)
 cargo +nightly fuzz cmin --fuzz-dir crates/fuzztau wire  crates/fuzztau/corpus/wire
@@ -148,4 +149,4 @@ cargo +nightly fuzz run --fuzz-dir crates/fuzztau wire crates/fuzztau/artifacts/
 | Unit tests | Regressions on known-shape behaviour | Always (CI) |
 | Hegel PBT | Invariant violations across random inputs | Always (inline with unit tests) |
 | libdst / dst | SUT vs reference divergence under emergent workloads | CI (after nextest, before Docker) |
-| cargo-fuzz | Panics, crashes, and OOMs from adversarial bytes on the text parsers and the binary `.dat` decoder | Time-boxed every CI run (build-check all targets; run `parse`/`wire`/`disk_decode`), plus on demand |
+| cargo-fuzz | Panics, crashes, and OOMs from adversarial bytes on the text parsers | Time-boxed every CI run (build-check all targets; run `parse`/`wire`), plus on demand |
